Cleanup for contact lists
Cleaned up the code for choosing the list factory as well as making it
possible to query what factories are available.

Also made the contact list have ownership of its handle, ensuring that
if the name is requested twice, it gets the same handle and channel
regardless of whether the client requests a hold.  telepathy-python
might have already been doing it, but oh well.